+++ src/main/java/org/tinspin/index/qthypercube2/QuadTreeKD2.java
@@ -42,7 +42,7 @@ import org.tinspin.index.util.StringBuilderLn;
  * With version 1, there were many nodes with just one data entry because the parent directory 
  * node could not hold date entries. 
  * With version two a directory node (which contains a hypercube array) will only
- * create a subnode if a quadrant has to mhold more than one data entry.
+ * create a subnode if a quadrant has to hold more than one data entry.
  * 
  * 
  * @author ztilmann
@@ -378,6 +378,7 @@ public class QuadTreeKD2<T> implements PointMap<T>, PointMultimap<T> {
 							  int depth, int posInParent) {
 		String prefix = ".".repeat(depth);
 		sb.append(prefix + posInParent + " d=" + depth);
+		sb.append(" nV=" + node.getValueCount());
 		sb.append(" " + Arrays.toString(node.getCenter()));
 		sb.appendLn("/" + node.getRadius());
 		prefix += " ";
@@ -418,7 +419,7 @@ public class QuadTreeKD2<T> implements PointMap<T>, PointMultimap<T> {
 	 * Statistics container class.
 	 */
 	public static class QStats extends Stats {
-		final int[] histoValues = new int[100];
+		final int[] histoValues = new int[1000];
 		final int[] histoSubs;
 		static final int HISTO_MAX = (1 << 10) + 1;
 		public QStats(int dims) {